\name{FinalReport2}
\alias{FinalReport2}
\docType{data}
\title{
  Allele call and intensity data for the second set of MUGA samples.
}
\description{
  Allele call and intensity data for the second set of MUGA samples.
}
\usage{data(FinalReport2)}
\format{
  Text with tab delimeters.
}
\details{
  This is the *_FinalReport.txt file as delivered by GeneSeek. It is unparsed text that 
  contains tab characters and is intended to be written out at text and then read back
  in to simulate the MUGA genotyping pipeline as the user would experience it.
}
\source{
  High-resolution genetic mapping using the Mouse Diversity outbred population.
  Svenson KL, Gatti DM, Valdar W, Welsh CE, Cheng R, Chesler EJ, Palmer AA, McMillan L, Churchill GA.
  Genetics. 2012 Feb;190(2):437-47. doi: 10.1534/genetics.111.132597.
  PMID: 22345611
}
\examples{
  data(FinalReport2)
}
\keyword{datasets}
