\name{MUGAExampleData}
\alias{MUGAExampleData}
\alias{MUGAExampleData}
\docType{package}
\title{
  Example {M}ouse {U}niversal {G}enotyping {A}rray data for genome reconstruction and quantitative trait locus mapping.
}
\description{
  Example MUGA data for a set of ~150 samples from Svenson et.al., Genetics, 2012.
}
\details{
\tabular{ll}{
  Package: \tab MUGAExampleData\cr
  Type: \tab Package\cr
  Version: \tab 1.0\cr
  Date: \tab 2013-10-02\cr
  License: \tab What license is it under?\cr
}

}
\author{
Daniel Gatti <Dan.Gatti@jax.org>
}
\references{
  High-resolution genetic mapping using the Mouse Diversity outbred population.
  Svenson KL, Gatti DM, Valdar W, Welsh CE, Cheng R, Chesler EJ, Palmer AA, McMillan L, Churchill GA.
  Genetics. 2012 Feb;190(2):437-47. doi: 10.1534/genetics.111.132597.
  PMID: 22345611
}
\keyword{ MUGA }
\examples{
  data(FinalReport1)
  data(Samples1)
}
