\name{pheno}
\alias{pheno}
\docType{data}
\title{
  Phenotype data for Svenson et.al samples.
}
\description{
  The clinical phenotypes for each mouse in the Svenson et.al. data set.
}
\usage{data(pheno)}
\format{
   Data.frame containing sample IDs, sex, diet and measured clinical phenotypes for 149 DO mice. 
}
\details{
  The clinical phenotypes for each mouse in the Svenson et.al. data set.
}
\source{
  High-resolution genetic mapping using the Mouse Diversity outbred population.
  Svenson KL, Gatti DM, Valdar W, Welsh CE, Cheng R, Chesler EJ, Palmer AA, McMillan L, Churchill GA.
  Genetics. 2012 Feb;190(2):437-47. doi: 10.1534/genetics.111.132597.
  PMID: 22345611
}
\examples{
  data(pheno)
}
\keyword{datasets}
