\name{Samples1}
\alias{Samples1}
\docType{data}
\title{
  Sample list for the first MUGA data set.
}
\description{
  List of samples that correspond to the samples in FinalReport1.
}
\usage{data(Samples1)}
\format{
   Table with tab delimiters.
}
\details{
  This is the Sample_Map.txt file as delivered by GeneSeek. It is unparsed text that 
  contains tab characters and is intended to be written out at text and then read back
  in to simulate the MUGA genotyping pipeline as the user would experience it.
}
\source{
  High-resolution genetic mapping using the Mouse Diversity outbred population.
  Svenson KL, Gatti DM, Valdar W, Welsh CE, Cheng R, Chesler EJ, Palmer AA, McMillan L, Churchill GA.
  Genetics. 2012 Feb;190(2):437-47. doi: 10.1534/genetics.111.132597.
  PMID: 22345611
}
\examples{
  data(Samples1)
}
\keyword{datasets}
