\name{x}
\alias{x}
\docType{data}
\title{
  X allele intensities for Svenson et.al samples.
}
\description{
  Matrix of allele intensities from the MUGA for each of the Svenson et.al. samples at each marker.
}
\usage{data(x)}
\format{
   X allele intensities for each sample at each marker. num.samples x num.snps. Sample names in rownames and SNP IDs in colnames.
}
\details{
  This is the summary of the X allele intensities from each of the raw data files (FinalReport1 and FinalReport2).
}
\source{
  High-resolution genetic mapping using the Mouse Diversity outbred population.
  Svenson KL, Gatti DM, Valdar W, Welsh CE, Cheng R, Chesler EJ, Palmer AA, McMillan L, Churchill GA.
  Genetics. 2012 Feb;190(2):437-47. doi: 10.1534/genetics.111.132597.
  PMID: 22345611
}
\examples{
  data(x)
}
\keyword{datasets}
